Renamed and reorged the file structure
Deleted unused files
