--- app/src/main/java/com/brainwallet/ui/screens/settings/BentoRailLearningBanner.kt
@@ -1,58 +1,0 @@
-package com.brainwallet.ui.screens.settings
-
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.PreviewLightDark
-import androidx.compose.ui.unit.dp
-import com.brainwallet.ui.theme.DesignTheme
-
-@Composable
-fun BentoRailLearningBanner(
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit = {}
-) {
-    Card(
-        modifier = modifier.fillMaxWidth(),
-        shape = RoundedCornerShape(16.dp),
-        elevation = CardDefaults.cardElevation(defaultElevation = 0.dp),
-        colors = CardDefaults.cardColors(
-            containerColor = DesignTheme.colors.background.copy(alpha = 0.2f)
-        ),
-        onClick = onClick
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f),
-            contentAlignment = Alignment.Center
-        ) {
-            Text(
-                "Learning Banner",
-                style = DesignTheme.typography.bodyLarge.copy(
-                    color = DesignTheme.colors.content,
-                    textAlign = TextAlign.Center,
-                    fontWeight = FontWeight.Bold
-                ),
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-    }
-}
-
-@Composable
-@PreviewLightDark
-fun BentoRailLearningBannerPreview() {
-    DesignTheme(isSystemInDarkTheme()) {
-        BentoRailLearningBanner()
-    }
-}
